requirements.txt
mypy
flake8
numpy
opencv-python
pillow
pylint
autopep8